feat: check if paths are outside snap restrictions
The flamenco snap runs in strict confinement and
can only access files inside the /home directory
(see https://snapcraft.io/docs/home-interface).

This commit adds error message to make users aware
of the true cause, when commands fail.

For example, if the user specifies to build inside
the /tmp directory, flamenco will build inside the
snap confinement and my report success, even
though the build output is not findable for the
user in /tmp.
Because of the snap confinement /tmp from the
perspective of a snap is not /tmp from perspective
of a user.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>