feat: add git external source support
This commit adds Git external source support with `GitExternalSource`. The implementation is made extensible by adding an `ExternalSourceBase` as an abstract base class that creates external source objects based on the type parameter of the source descriptors.

The git manipulation code takes a dependency on the `LibGit2Sharp` NuGet package.