feat: cache Git repository to avoid multiple clones
This code makes use of a dictionary of SemaphoreSlim
objects to orchestrate the cloning of individual
repositories.

While a repository is cloning, all operations for that
repository are blocked. When finished, all subsequent
operations on that repository will be copies from the
cache directory to its final destination.

This will prevent the same repository being cloned
`release` x `package` times.