feat: add support for ignored files in Git external source
The `ignoredFiles` JSON key will contain a list of files in the repository that should be deleted after cloning. This is useful for scenarios where the user does not wish to keep certain files around when cloning a repository into a Debian package, e.g. .gitignore files.

Files that are listed and do not exist in the repository are ignored.